Added 'Model View Screen Modifier Key' and various improvements

+++ common/src/main/java/com/xtracr/realcamera/util/SmoothUtil.java
@@ -19,7 +19,7 @@ public class SmoothUtil {
     }
 
     public static Quaterniond smoothRotation(Quaterniond rotation) {
-        lastRotation.slerp(rotation, 1 - (float) ConfigFile.config().getRotationSmoothFactor());
+        lastRotation.slerp(rotation, 1 - ConfigFile.config().getRotationSmoothFactor());
         return lastRotation;
     }
 }
